app/client/code/requirements.txt
click==8.1.7
Flask==3.1.0
itsdangerous==2.1.0
Jinja2==3.1.5
MarkupSafe==3.0.1
psycopg2-binary=2.9.1
protobuf==5.29.1
six==1.17.0
Werkzeug==3.1.1
